Split out project templates test projects to reduce Helix job run time
- Made launch profile checks more stringent
