+++ src/ProjectTemplates/ProjectTemplates.slnf
@@ -67,6 +67,7 @@
       "src\\ProjectTemplates\\Web.ItemTemplates\\Microsoft.DotNet.Web.ItemTemplates.csproj",
       "src\\ProjectTemplates\\Web.ProjectTemplates\\Microsoft.DotNet.Web.ProjectTemplates.csproj",
       "src\\ProjectTemplates\\test\\ProjectTemplates.Tests\\ProjectTemplates.Tests.csproj",
+      "src\\ProjectTemplates\\test\\ProjectTemplates.Mvc.Tests\\ProjectTemplates.Mvc.Tests.csproj",
       "src\\Razor\\Razor.Runtime\\src\\Microsoft.AspNetCore.Razor.Runtime.csproj",
       "src\\Razor\\Razor\\src\\Microsoft.AspNetCore.Razor.csproj",
       "src\\Security\\Authentication\\Cookies\\src\\Microsoft.AspNetCore.Authentication.Cookies.csproj",
