+++ src/ProjectTemplates/Shared/ArgConstants.cs
@@ -1,7 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace Templates.Test;
+namespace Templates.Test.Helpers;
 
 internal static class ArgConstants
 {
