+++ src/ProjectTemplates/test/ProjectTemplates.Blazor.Tests/.gitattributes
@@ -1,0 +1,1 @@
+.json diff
