+++ src/ProjectTemplates/test/ProjectTemplates.Blazor.Tests/BlazorWasmTemplateTest.cs
@@ -15,7 +15,7 @@ using Xunit;
 using Xunit.Abstractions;
 using Xunit.Sdk;
 
-namespace Templates.Test;
+namespace Templates.Blazor.Test;
 
 public class BlazorWasmTemplateTest : BlazorTemplateTest
 {
