+++ src/ProjectTemplates/test/ProjectTemplates.Mvc.Tests/.gitattributes
@@ -1,0 +1,1 @@
+.json diff
