+++ src/ProjectTemplates/test/ProjectTemplates.Mvc.Tests/RazorPagesTemplateTest.cs
@@ -1,20 +1,25 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Testing;
 using Templates.Test.Helpers;
+using Xunit;
 using Xunit.Abstractions;
 
-namespace Templates.Test;
+namespace Templates.Mvc.Test;
 
-public class MvcTemplateTest : LoggedTest
+public class RazorPagesTemplateTest : LoggedTest
 {
-    public MvcTemplateTest(ProjectFactoryFixture projectFactory)
+    public RazorPagesTemplateTest(ProjectFactoryFixture projectFactory)
     {
         ProjectFactory = projectFactory;
     }
 
-    public ProjectFactoryFixture ProjectFactory { get; }
+    public ProjectFactoryFixture ProjectFactory { get; set; }
 
     private ITestOutputHelper _output;
     public ITestOutputHelper Output
@@ -29,86 +34,71 @@ public class MvcTemplateTest : LoggedTest
         }
     }
 
-    [Fact]
-    public async Task MvcTemplate_NoAuthFSharp() => await MvcTemplateCore(languageOverride: "F#");
-
-    [Fact]
-    public async Task MvcTemplate_NoAuthNoHttpsFSharp() => await MvcTemplateCore(languageOverride: "F#", args: new[] { ArgConstants.NoHttps } );
-
-    [ConditionalFact]
-    [SkipOnHelix("Cert failure, https://github.com/dotnet/aspnetcore/issues/28090", Queues = "All.OSX;" + HelixConstants.Windows10Arm64 + HelixConstants.DebianArm64 + HelixConstants.DebianAmd64)]
-    public async Task MvcTemplate_NoAuthCSharp() => await MvcTemplateCore(languageOverride: null);
-
-    [ConditionalFact]
-    [SkipOnHelix("Cert failure, https://github.com/dotnet/aspnetcore/issues/28090", Queues = "All.OSX;" + HelixConstants.Windows10Arm64 + HelixConstants.DebianArm64 + HelixConstants.DebianAmd64)]
-    public async Task MvcTemplate_NoAuthNoHttpsCSharp() => await MvcTemplateCore(languageOverride: null, new[] { ArgConstants.NoHttps });
-
-    [ConditionalFact]
-    [SkipOnHelix("Cert failure, https://github.com/dotnet/aspnetcore/issues/28090", Queues = "All.OSX;" + HelixConstants.Windows10Arm64 + HelixConstants.DebianArm64 + HelixConstants.DebianAmd64)]
-    public async Task MvcTemplate_ProgramMainNoAuthCSharp() => await MvcTemplateCore(languageOverride: null, new[] { ArgConstants.UseProgramMain });
-
-    [ConditionalFact]
-    [SkipOnHelix("Cert failure, https://github.com/dotnet/aspnetcore/issues/28090", Queues = "All.OSX;" + HelixConstants.Windows10Arm64 + HelixConstants.DebianArm64 + HelixConstants.DebianAmd64)]
-    public async Task MvcTemplate_ProgramMainNoAuthNoHttpsCSharp() => await MvcTemplateCore(languageOverride: null, new[] { ArgConstants.UseProgramMain, ArgConstants.NoHttps });
-
-    private async Task MvcTemplateCore(string languageOverride, string[] args = null)
+    [ConditionalTheory]
+    [SkipOnHelix("Cert failure, https://github.com/dotnet/aspnetcore/issues/28090", Queues = "All.OSX;" + HelixConstants.Windows10Arm64 + HelixConstants.DebianArm64)]
+    [InlineData(true, false)]
+    [InlineData(true, true)]
+    [InlineData(false, false)]
+    [InlineData(false, true)]
+    public async Task RazorPagesTemplate_NoAuth(bool useProgramMain, bool noHttps)
     {
         var project = await ProjectFactory.CreateProject(Output);
 
-        var createResult = await project.RunDotNetNewAsync("mvc", language: languageOverride, args: args);
-        Assert.True(0 == createResult.ExitCode, ErrorMessages.GetFailedProcessMessage("create/restore", project, createResult));
+        var args = useProgramMain
+            ? noHttps
+                ? new[] { ArgConstants.UseProgramMain, ArgConstants.NoHttps }
+                : new[] { ArgConstants.UseProgramMain }
+            : noHttps
+                ? new[] { ArgConstants.NoHttps }
+                : null;
+        var createResult = await project.RunDotNetNewAsync("razor", args: args);
+        Assert.True(0 == createResult.ExitCode, ErrorMessages.GetFailedProcessMessage("razor", project, createResult));
 
-        var noHttps = args?.Contains(ArgConstants.NoHttps) ?? false;
         var expectedLaunchProfileNames = noHttps
             ? new[] { "http", "IIS Express" }
             : new[] { "http", "https", "IIS Express" };
         await project.VerifyLaunchSettings(expectedLaunchProfileNames);
 
-        var projectExtension = languageOverride == "F#" ? "fsproj" : "csproj";
-        var projectFileContents = project.ReadFile($"{project.ProjectName}.{projectExtension}");
+        var projectFileContents = ReadFile(project.TemplateOutputDir, $"{project.ProjectName}.csproj");
         Assert.DoesNotContain(".db", projectFileContents);
         Assert.DoesNotContain("Microsoft.EntityFrameworkCore.Tools", projectFileContents);
         Assert.DoesNotContain("Microsoft.VisualStudio.Web.CodeGeneration.Design", projectFileContents);
         Assert.DoesNotContain("Microsoft.EntityFrameworkCore.Tools.DotNet", projectFileContents);
         Assert.DoesNotContain("Microsoft.Extensions.SecretManager.Tools", projectFileContents);
 
-        // Avoid the F# compiler. See https://github.com/dotnet/aspnetcore/issues/14022
-        if (languageOverride != null)
-        {
-            return;
-        }
-
         var publishResult = await project.RunDotNetPublishAsync();
-        Assert.True(0 == publishResult.ExitCode, ErrorMessages.GetFailedProcessMessage("publish", project, publishResult));
+        Assert.True(0 == publishResult.ExitCode, ErrorMessages.GetFailedProcessMessage("publish", project, createResult));
 
         // Run dotnet build after publish. The reason is that one uses Config = Debug and the other uses Config = Release
         // The output from publish will go into bin/Release/netcoreappX.Y/publish and won't be affected by calling build
         // later, while the opposite is not true.
 
         var buildResult = await project.RunDotNetBuildAsync();
-        Assert.True(0 == buildResult.ExitCode, ErrorMessages.GetFailedProcessMessage("build", project, buildResult));
-
-        IEnumerable<string> menuLinks = new List<string> {
-                PageUrls.HomeUrl,
-                PageUrls.HomeUrl,
-                PageUrls.PrivacyFullUrl
-            };
-
-        var footerLinks = new string[] { PageUrls.PrivacyFullUrl };
+        Assert.True(0 == buildResult.ExitCode, ErrorMessages.GetFailedProcessMessage("build", project, createResult));
 
         var pages = new List<Page>
-        {
-            new Page
             {
-                Url = PageUrls.HomeUrl,
-                Links = menuLinks.Append(PageUrls.DocsUrl).Concat(footerLinks)
-            },
-            new Page
-            {
-                Url = PageUrls.PrivacyFullUrl,
-                Links = menuLinks.Concat(footerLinks)
-            }
-        };
+                new Page
+                {
+                    Url = PageUrls.HomeUrl,
+                    Links = new [] {
+                        PageUrls.HomeUrl,
+                        PageUrls.HomeUrl,
+                        PageUrls.PrivacyUrl,
+                        PageUrls.DocsUrl,
+                        PageUrls.PrivacyUrl
+                    }
+                },
+                new Page
+                {
+                    Url = PageUrls.PrivacyUrl,
+                    Links = new [] {
+                        PageUrls.HomeUrl,
+                        PageUrls.HomeUrl,
+                        PageUrls.PrivacyUrl,
+                        PageUrls.PrivacyUrl }
+                }
+            };
 
         using (var aspNetProcess = project.StartBuiltProjectAsync())
         {
@@ -136,7 +126,7 @@ public class MvcTemplateTest : LoggedTest
     [InlineData(true, true)]
     [SkipOnHelix("Cert failure, https://github.com/dotnet/aspnetcore/issues/28090", Queues = "All.OSX;" + HelixConstants.Windows10Arm64 + HelixConstants.DebianArm64 + HelixConstants.DebianAmd64)]
     [OSSkipCondition(OperatingSystems.Linux | OperatingSystems.MacOSX, SkipReason = "No LocalDb on non-Windows")]
-    public Task MvcTemplate_IndividualAuth_LocalDb(bool useProgramMain, bool noHttps) => MvcTemplate_IndividualAuth_Core(useLocalDB: true, useProgramMain, noHttps);
+    public Task RazorPagesTemplate_IndividualAuth_LocalDb(bool useProgramMain, bool noHttps) => RazorPagesTemplate_IndividualAuth_Core(useLocalDB: true, useProgramMain, noHttps);
 
     [ConditionalTheory]
     [InlineData(false, false)]
@@ -144,9 +134,9 @@ public class MvcTemplateTest : LoggedTest
     [InlineData(true, false)]
     [InlineData(true, true)]
     [SkipOnHelix("Cert failure, https://github.com/dotnet/aspnetcore/issues/28090", Queues = "All.OSX;" + HelixConstants.Windows10Arm64 + HelixConstants.DebianArm64 + HelixConstants.DebianAmd64)]
-    public Task MvcTemplate_IndividualAuth(bool useProgramMain, bool noHttps) => MvcTemplate_IndividualAuth_Core(useLocalDB: false, useProgramMain, noHttps);
+    public Task RazorPagesTemplate_IndividualAuth(bool useProgramMain, bool noHttps) => RazorPagesTemplate_IndividualAuth_Core(useLocalDB: false, useProgramMain, noHttps);
 
-    private async Task MvcTemplate_IndividualAuth_Core(bool useLocalDB, bool useProgramMain, bool noHttps)
+    private async Task RazorPagesTemplate_IndividualAuth_Core(bool useLocalDB, bool useProgramMain, bool noHttps)
     {
         var project = await ProjectFactory.CreateProject(Output);
 
@@ -157,15 +147,16 @@ public class MvcTemplateTest : LoggedTest
             : noHttps
                 ? new[] { ArgConstants.NoHttps }
                 : null;
-        var createResult = await project.RunDotNetNewAsync("mvc", auth: "Individual", useLocalDB: useLocalDB, args: args);
+        var createResult = await project.RunDotNetNewAsync("razor", auth: "Individual", useLocalDB: useLocalDB, args: args);
         Assert.True(0 == createResult.ExitCode, ErrorMessages.GetFailedProcessMessage("create/restore", project, createResult));
 
+        // Individual auth supports no https OK
         var expectedLaunchProfileNames = noHttps
             ? new[] { "http", "IIS Express" }
             : new[] { "http", "https", "IIS Express" };
         await project.VerifyLaunchSettings(expectedLaunchProfileNames);
 
-        var projectFileContents = project.ReadFile($"{project.ProjectName}.csproj");
+        var projectFileContents = ReadFile(project.TemplateOutputDir, $"{project.ProjectName}.csproj");
         if (!useLocalDB)
         {
             Assert.Contains(".db", projectFileContents);
@@ -181,11 +172,11 @@ public class MvcTemplateTest : LoggedTest
         var buildResult = await project.RunDotNetBuildAsync();
         Assert.True(0 == buildResult.ExitCode, ErrorMessages.GetFailedProcessMessage("build", project, buildResult));
 
-        var migrationsResult = await project.RunDotNetEfCreateMigrationAsync("mvc");
+        var migrationsResult = await project.RunDotNetEfCreateMigrationAsync("razorpages");
         Assert.True(0 == migrationsResult.ExitCode, ErrorMessages.GetFailedProcessMessage("run EF migrations", project, migrationsResult));
-        project.AssertEmptyMigration("mvc");
+        project.AssertEmptyMigration("razorpages");
 
-        // Note: if any links are updated here, RazorPagesTemplateTest.cs should be updated as well
+        // Note: if any links are updated here, MvcTemplateTest.cs should be updated as well
         var pages = new List<Page> {
                 new Page
                 {
@@ -214,7 +205,7 @@ public class MvcTemplateTest : LoggedTest
                 },
                 new Page
                 {
-                    Url = PageUrls.PrivacyFullUrl,
+                    Url = PageUrls.PrivacyUrl,
                     Links = new [] {
                         PageUrls.HomeUrl,
                         PageUrls.HomeUrl,
@@ -267,101 +258,46 @@ public class MvcTemplateTest : LoggedTest
         {
             Assert.False(
                 aspNetProcess.Process.HasExited,
-                ErrorMessages.GetFailedProcessMessageOrEmpty("Run published project", project, aspNetProcess.Process));
+                ErrorMessages.GetFailedProcessMessageOrEmpty("Run built project", project, aspNetProcess.Process));
 
             await aspNetProcess.AssertPagesOk(pages);
         }
-    }
-
-    [ConditionalFact]
-    [OSSkipCondition(OperatingSystems.Linux | OperatingSystems.MacOSX)] // Running these requires the rid-specific runtime pack to be available which is not consistent in all our platform builds.
-    [SkipOnHelix("cert failure", Queues = "All.OSX;" + HelixConstants.Windows10Arm64)]
-    public async Task MvcTemplate_SingleFileExe()
-    {
-        // This test verifies publishing an MVC app as a single file exe works. We'll limit testing
-        // this to a few operating systems to make our lives easier.
-        var runtimeIdentifer = "win-x64";
-        var project = await ProjectFactory.CreateProject(Output);
-        project.RuntimeIdentifier = runtimeIdentifer;
-
-        var createResult = await project.RunDotNetNewAsync("mvc");
-        Assert.True(0 == createResult.ExitCode, ErrorMessages.GetFailedProcessMessage("create/restore", project, createResult));
-
-        var publishResult = await project.RunDotNetPublishAsync(additionalArgs: $"/p:PublishSingleFile=true -r {runtimeIdentifer} --self-contained", noRestore: false);
-        Assert.True(0 == publishResult.ExitCode, ErrorMessages.GetFailedProcessMessage("publish", project, publishResult));
-
-        var menuLinks = new[]
-        {
-            PageUrls.HomeUrl,
-            PageUrls.HomeUrl,
-            PageUrls.PrivacyFullUrl
-        };
-
-        var footerLinks = new[] { PageUrls.PrivacyFullUrl };
-
-        var pages = new List<Page>
-        {
-            new Page
-            {
-                Url = PageUrls.HomeUrl,
-                Links = menuLinks.Append(PageUrls.DocsUrl).Concat(footerLinks),
-            },
-            new Page
-            {
-                Url = PageUrls.PrivacyFullUrl,
-                Links = menuLinks.Concat(footerLinks),
-            }
-        };
-
-        using var aspNetProcess = project.StartPublishedProjectAsync(usePublishedAppHost: true);
-        Assert.False(
-            aspNetProcess.Process.HasExited,
-            ErrorMessages.GetFailedProcessMessageOrEmpty("Run published project", project, aspNetProcess.Process));
-
-        await aspNetProcess.AssertPagesOk(pages);
     }
 
     [ConditionalTheory]
     [SkipOnHelix("https://github.com/dotnet/aspnetcore/issues/28090", Queues = HelixConstants.Windows10Arm64 + HelixConstants.DebianArm64)]
     [InlineData("IndividualB2C", null)]
     [InlineData("IndividualB2C", new[] { ArgConstants.UseProgramMain })]
-    [InlineData("IndividualB2C", new[] { ArgConstants.CalledApiUrlGraphMicrosoftCom, ArgConstants.CalledApiScopesUserReadWrite })]
-    [InlineData("IndividualB2C", new[] { ArgConstants.UseProgramMain, ArgConstants.CalledApiUrlGraphMicrosoftCom, ArgConstants.CalledApiScopesUserReadWrite })]
-    public Task MvcTemplate_IdentityWeb_IndividualB2C_BuildsAndPublishes(string auth, string[] args) => MvcTemplateBuildsAndPublishes(auth: auth, args: args);
-
-    [ConditionalTheory]
-    [SkipOnHelix("https://github.com/dotnet/aspnetcore/issues/28090", Queues = HelixConstants.Windows10Arm64 + HelixConstants.DebianArm64)]
-    [InlineData("IndividualB2C", null)]
     [InlineData("IndividualB2C", new[] { ArgConstants.UseProgramMain, ArgConstants.NoHttps })]
+    [InlineData("IndividualB2C", new[] { ArgConstants.CalledApiUrlGraphMicrosoftCom, ArgConstants.CalledApiScopesUserReadWrite })]
     [InlineData("IndividualB2C", new[] { ArgConstants.CalledApiUrlGraphMicrosoftCom, ArgConstants.CalledApiScopesUserReadWrite, ArgConstants.NoHttps })]
+    [InlineData("IndividualB2C", new[] { ArgConstants.UseProgramMain, ArgConstants.CalledApiUrlGraphMicrosoftCom, ArgConstants.CalledApiScopesUserReadWrite })]
     [InlineData("IndividualB2C", new[] { ArgConstants.UseProgramMain, ArgConstants.CalledApiUrlGraphMicrosoftCom, ArgConstants.CalledApiScopesUserReadWrite, ArgConstants.NoHttps })]
-    public Task MvcTemplate_IdentityWeb_IndividualB2C_NoHttps_BuildsAndPublishes(string auth, string[] args) => MvcTemplateBuildsAndPublishes(auth: auth, args: args);
+    public Task RazorPagesTemplate_IdentityWeb_IndividualB2C_BuildsAndPublishes(string auth, string[] args) => BuildAndPublishRazorPagesTemplateIdentityWeb(auth: auth, args: args);
 
     [ConditionalTheory]
     [SkipOnHelix("https://github.com/dotnet/aspnetcore/issues/28090", Queues = HelixConstants.Windows10Arm64 + HelixConstants.DebianArm64)]
     [InlineData("SingleOrg", null)]
     [InlineData("SingleOrg", new[] { ArgConstants.UseProgramMain })]
+    [InlineData("SingleOrg", new[] { ArgConstants.UseProgramMain, ArgConstants.NoHttps })]
     [InlineData("SingleOrg", new[] { ArgConstants.CalledApiUrlGraphMicrosoftCom, ArgConstants.CalledApiScopesUserReadWrite })]
+    [InlineData("SingleOrg", new[] { ArgConstants.CalledApiUrlGraphMicrosoftCom, ArgConstants.CalledApiScopesUserReadWrite, ArgConstants.NoHttps })]
     [InlineData("SingleOrg", new[] { ArgConstants.UseProgramMain, ArgConstants.CalledApiUrlGraphMicrosoftCom, ArgConstants.CalledApiScopesUserReadWrite })]
-    [InlineData("SingleOrg", new[] { ArgConstants.CallsGraph })]
-    [InlineData("SingleOrg", new[] { ArgConstants.UseProgramMain, ArgConstants.CallsGraph })]
-    public Task MvcTemplate_IdentityWeb_SingleOrg_BuildsAndPublishes(string auth, string[] args) => MvcTemplateBuildsAndPublishes(auth: auth, args: args);
+    [InlineData("SingleOrg", new[] { ArgConstants.UseProgramMain, ArgConstants.CalledApiUrlGraphMicrosoftCom, ArgConstants.CalledApiScopesUserReadWrite, ArgConstants.NoHttps })]
+    public Task RazorPagesTemplate_IdentityWeb_SingleOrg_BuildsAndPublishes(string auth, string[] args) => BuildAndPublishRazorPagesTemplateIdentityWeb(auth: auth, args: args);
 
     [ConditionalTheory]
-    [SkipOnHelix("https://github.com/dotnet/aspnetcore/issues/28090", Queues = HelixConstants.Windows10Arm64 + HelixConstants.DebianArm64)]
-    [InlineData("SingleOrg", null)]
-    [InlineData("SingleOrg", new[] { ArgConstants.UseProgramMain, ArgConstants.NoHttps })]
-    [InlineData("SingleOrg", new[] { ArgConstants.CalledApiUrlGraphMicrosoftCom, ArgConstants.CalledApiScopesUserReadWrite, ArgConstants.NoHttps })]
-    [InlineData("SingleOrg", new[] { ArgConstants.UseProgramMain, ArgConstants.CalledApiUrlGraphMicrosoftCom, ArgConstants.CalledApiScopesUserReadWrite, ArgConstants.NoHttps })]
+    [InlineData("SingleOrg", new[] { ArgConstants.CallsGraph })]
+    [InlineData("SingleOrg", new[] { ArgConstants.UseProgramMain, ArgConstants.CallsGraph })]
     [InlineData("SingleOrg", new[] { ArgConstants.CallsGraph, ArgConstants.NoHttps })]
     [InlineData("SingleOrg", new[] { ArgConstants.UseProgramMain, ArgConstants.CallsGraph, ArgConstants.NoHttps })]
-    public Task MvcTemplate_IdentityWeb_SingleOrg_NoHttps_BuildsAndPublishes(string auth, string[] args) => MvcTemplateBuildsAndPublishes(auth: auth, args: args);
+    public Task RazorPagesTemplate_IdentityWeb_SingleOrg_CallsGraph_BuildsAndPublishes(string auth, string[] args) => BuildAndPublishRazorPagesTemplateIdentityWeb(auth: auth, args: args);
 
-    private async Task<Project> MvcTemplateBuildsAndPublishes(string auth, string[] args)
+    private async Task<Project> BuildAndPublishRazorPagesTemplateIdentityWeb(string auth, string[] args)
     {
         var project = await ProjectFactory.CreateProject(Output);
 
-        var createResult = await project.RunDotNetNewAsync("mvc", auth: auth, args: args);
+        var createResult = await project.RunDotNetNewAsync("razor", auth: auth, args: args);
         Assert.True(0 == createResult.ExitCode, ErrorMessages.GetFailedProcessMessage("create/restore", project, createResult));
 
         // Identity Web auth requires https and thus ignores the --no-https option if passed so there should never be an 'http' profile
@@ -377,5 +313,14 @@ public class MvcTemplateTest : LoggedTest
         Assert.True(0 == buildResult.ExitCode, ErrorMessages.GetFailedProcessMessage("publish", project, buildResult));
 
         return project;
+    }
+
+    private string ReadFile(string basePath, string path)
+    {
+        var fullPath = Path.Combine(basePath, path);
+        var doesExist = File.Exists(fullPath);
+
+        Assert.True(doesExist, $"Expected file to exist, but it doesn't: {path}");
+        return File.ReadAllText(Path.Combine(basePath, path));
     }
 }
