+++ src/ProjectTemplates/test/ProjectTemplates.Mvc.Tests/WebApiTemplateTest.cs
@@ -10,7 +10,7 @@ using Templates.Test.Helpers;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Templates.Test;
+namespace Templates.Mvc.Test;
 
 public class WebApiTemplateTest : LoggedTest
 {
